pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdt"
version = "0.1.0"
description = "Rambo Docker Tools — CLI for docker-compose generation"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12",
    "ruamel.yaml>=0.18",
    "questionary>=2.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "port-for>=0.7",
    "rich>=13.0",
]

[project.scripts]
rdt = "rdt.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["rdt*"]

